Basic CSV parsing
It can read out a CSV file and parse it, it just doesn't parse it right, the PGN defs from the DBC are still fucked so the log parse button only tests that to show you that it is in fact fucked. Also the structure of this is rapidly devloving and I need to clean up some of the structs for bg utils, also the proto-buf thing "works" it does encode and decode a protobuf packet, and that is about it.